devblogger/pyproject.toml
[project]
name = "devblogger"
version = "0.1.0"
description = "Semi-automatic development blog system for generating blog entries from GitHub commits"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "DevBlogger Team", email = "team@devblogger.dev"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
    "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
    "customtkinter>=5.2.0",
    "CTkMessagebox>=2.5",
    "CTkListbox>=1.0",
    "requests>=2.31.0",
    "aiohttp>=3.8.0",
    "google-generativeai>=0.3.0",
    "openai>=1.0.0",
    "requests-oauthlib>=1.3.0",
    "python-dotenv>=1.0.0",
    "markdown>=3.5.0",
    "pyinstaller>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[project.scripts]
devblogger = "src.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/devblogger"
Repository = "https://github.com/yourusername/devblogger"
Issues = "https://github.com/yourusername/devblogger/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
